theia-core/pyproject.toml
[project]
name = "theia-core"
version = "0.0.1"
description = "Hermes session JSON → theia graph.json"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "umap-learn>=0.5",
    "jsonschema>=4.21",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
theia-core = "theia_core.__main__:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
